Moving operator overrides into a Numeric class and using this for both Constant and Enum
