+++ packtype/primitive.py
@@ -20,6 +20,7 @@ except ImportError:
     from typing_extensions import Self  # noqa: UP035
 
 from .base import Base, MetaBase
+from .numeric import Numeric
 
 
 class PrimitiveValueError(Exception):
@@ -51,7 +52,7 @@ class MetaPrimitive(MetaBase):
         )
 
 
-class Primitive(Base, metaclass=MetaPrimitive):
+class Primitive(Base, Numeric, metaclass=MetaPrimitive):
     _PT_WIDTH: int = -1
     _PT_SIGNED: bool = False
 
@@ -94,157 +95,3 @@ class Primitive(Base, metaclass=MetaPrimitive):
 
     def __int__(self) -> int:
         return self.__value
-
-    # ==========================================================================
-    # Operator overrides
-    # ==========================================================================
-
-    def __add__(self, other: int | Self) -> int:
-        return int(self) + int(other)
-
-    def __sub__(self, other: int | Self) -> int:
-        return int(self) - int(other)
-
-    def __mul__(self, other: int | Self) -> int:
-        if isinstance(other, MetaBase):
-            return int(self) * other
-        else:
-            return int(self) * int(other)
-
-    def __truediv__(self, other: int | Self) -> int:
-        return int(self) / int(other)
-
-    def __floordiv__(self, other: int | Self) -> int:
-        return int(self) // int(other)
-
-    def __mod__(self, other: int | Self) -> int:
-        return int(self) % int(other)
-
-    def __divmod__(self, other: int | Self) -> int:
-        return (int(self) // int(other), int(self) % int(other))
-
-    def __pow__(self, other: int | Self) -> int:
-        return int(self) ** int(other)
-
-    def __lshift__(self, other: int | Self) -> int:
-        return int(self) << int(other)
-
-    def __rshift__(self, other: int | Self) -> int:
-        return int(self) >> int(other)
-
-    def __and__(self, other: int | Self) -> int:
-        return int(self) & int(other)
-
-    def __xor__(self, other: int | Self) -> int:
-        return int(self) ^ int(other)
-
-    def __or__(self, other: int | Self) -> int:
-        return int(self) | int(other)
-
-    def __radd__(self, other: int | Self) -> int:
-        return int(other) + int(self)
-
-    def __rsub__(self, other: int | Self) -> int:
-        return int(other) - int(self)
-
-    def __rmul__(self, other: int | Self) -> int:
-        return int(other) * int(self)
-
-    def __rtruediv__(self, other: int | Self) -> int:
-        return int(other) / int(self)
-
-    def __rfloordiv__(self, other: int | Self) -> int:
-        return int(other) // int(self)
-
-    def __rmod__(self, other: int | Self) -> int:
-        return int(other) % int(self)
-
-    def __rdivmod__(self, other: int | Self) -> int:
-        return (int(other) // int(self), int(other) % int(self))
-
-    def __rpow__(self, other: int | Self) -> int:
-        return int(other) ** int(self)
-
-    def __rlshift__(self, other: int | Self) -> int:
-        return int(other) << int(self)
-
-    def __rrshift__(self, other: int | Self) -> int:
-        return int(other) >> int(self)
-
-    def __rand__(self, other: int | Self) -> int:
-        return int(other) & int(self)
-
-    def __rxor__(self, other: int | Self) -> int:
-        return int(other) ^ int(self)
-
-    def __ror__(self, other: int | Self) -> int:
-        return int(other) | int(self)
-
-    def __iadd__(self, other: int | Self) -> int:
-        return int(self) + int(other)
-
-    def __isub__(self, other: int | Self) -> int:
-        return int(self) - int(other)
-
-    def __imul__(self, other: int | Self) -> int:
-        return int(self) * int(other)
-
-    def __itruediv__(self, other: int | Self) -> int:
-        return int(self) / int(other)
-
-    def __ifloordiv__(self, other: int | Self) -> int:
-        return int(self) // int(other)
-
-    def __imod__(self, other: int | Self) -> int:
-        return int(self) % int(other)
-
-    def __ipow__(self, other: int | Self) -> int:
-        return int(self) ** int(other)
-
-    def __ilshift__(self, other: int | Self) -> int:
-        return int(self) << int(other)
-
-    def __irshift__(self, other: int | Self) -> int:
-        return int(self) >> int(other)
-
-    def __iand__(self, other: int | Self) -> int:
-        return int(self) & int(other)
-
-    def __ixor__(self, other: int | Self) -> int:
-        return int(self) ^ int(other)
-
-    def __ior__(self, other: int | Self) -> int:
-        return int(self) | int(other)
-
-    def __neg__(self) -> int:
-        return -1 * int(self)
-
-    def __pos__(self) -> int:
-        return abs(int(self))
-
-    def __abs__(self) -> int:
-        return abs(int(self))
-
-    def __invert__(self) -> int:
-        return ~int(self)
-
-    def __lt__(self, other: int | Self) -> bool:
-        return int(self) < int(other)
-
-    def __le__(self, other: int | Self) -> bool:
-        return int(self) <= int(other)
-
-    def __eq__(self, other: int | Self) -> bool:
-        return int(self) == int(other)
-
-    def __ne__(self, other: int | Self) -> bool:
-        return int(self) != int(other)
-
-    def __gt__(self, other: int | Self) -> bool:
-        return int(self) > int(other)
-
-    def __ge__(self, other: int | Self) -> bool:
-        return int(self) >= int(other)
-
-    def __hash__(self) -> int:
-        return id(self)
